Continue work on refactoring App index.jsx
Have to get rid of local state and refactor Redux
store to have to separate reducers: one for
trending results (that includes the gifData, the
offset, totalCount, and scrolling bool) and the
other for search results. App index.jsx should
have access to both of these types of results
with all of the aforementioned properties/values.